skills/web_tools.py
import requests
from bs4 import BeautifulSoup
import re

class WebTools:
    @staticmethod
    def read_url(url: str):
        """
        Fetches and reads the textual content of a URL.
        Returns a simplified text representation of the page.
        """
        try:
            import urllib3
            urllib3.disable_warnings(urllib3.exceptions.InsecureRequestWarning)
            
            headers = {
                'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/91.0.4472.124 Safari/537.36'
            }
            response = requests.get(url, headers=headers, timeout=10, verify=False)
            response.raise_for_status()

            soup = BeautifulSoup(response.content, 'html.parser')

            # Kill all script and style elements
            for script in soup(["script", "style", "nav", "footer", "header", "noscript"]):
                script.decompose()

            # Get text
            text = soup.get_text()

            # Break into lines and remove leading/trailing space on each
            lines = (line.strip() for line in text.splitlines())
            # Break multi-headlines into a line each
            chunks = (phrase.strip() for line in lines for phrase in line.split("  "))
            # Drop blank lines
            text = '\n'.join(chunk for chunk in chunks if chunk)

            # Limit to a reasonable length to avoid context overflow (approx 4000 chars)
            if len(text) > 4000:
                text = text[:4000] + "\n...[Content Truncated]..."

            return f"--- Content of {url} ---\n{text}"

        except Exception as e:
            return f"Error reading URL '{url}': {str(e)}"
