requirements.txt
rich
langchain
langchain-community
langchain-core
python-dotenv
typer
openai
colorama
duckduckgo-search
dotenv
langchain-groq
ddgs
langchain_google_genai
requests
beautifulsoup4
pycaw
comtypes
psutil
pyperclip
Pillow
playwright
playwright-stealth
pywinauto
pyautogui
opencv-python
